Scmp clean up
